+++ src/constants/localStorage.js
@@ -1,0 +1,3 @@
+export const CUSTOM_QUERIES_LS = 'gv-custom=queries'
+export const STYLING_LS = 'gv-styling'
+export const GRAPH_VERSIONS_LS = 'gv-graph-versions'