+++ CHANGELOG.md
@@ -18,6 +18,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [0.3.0]
 ### Added
+- ONTOLOGY-20 Edges styling by property
+- ONTOLOGY-19 Nodes styling by property
 - ONTOLOGY-19 Functions  for VisJs Dataset manipulation
 - ONTOLOGY-20 Edges styling
 - ONTOLOGY-19 Nodes styling